Fix permissions issue preventing MLflow from saving models
Previously, MLflow was unable to save models due to a permission denied error. The issue was caused by the lack of proper permissions for the user running the MLflow container. To fix this, I added a `jovyan` user with the correct permissions and made sure that the user had access to the necessary directories. This allowed me to successfully save models from the MLflow UI and notebook.

This commit fixes the permission issue by updating the Dockerfile and creating a `jovyan` user with the necessary permissions.